CGFloat Convert

+++ Sources/LivenessCamera/View/LivenessCameraView.swift
@@ -79,7 +79,7 @@ public class LivenessCameraViewController: UIViewController, AVCapturePhotoCaptu
     private func animateProgress(progress: Int) {
         let basicAnimation = CABasicAnimation(keyPath: "strokeEnd")
         print("PROGRESS TAKEN \(progreeTaken)")
-        progreeTaken += counter / 15.0
+        progreeTaken += CGFloat(counter / 15.0)
 //        basicAnimation.toValue =  progreeTaken
         shape.strokeEnd = progreeTaken
         basicAnimation.duration = 2
